UP last cases with verbosity, checked for now simple cases left
The only remaining ifExists are:
* in src/
* plugins which don't have a cfi.py file
* plugins which do have, but the remaining parameters are either
  + alike to inputFiles
  + or a PSet/MappingEntry
* existsAs are used only for PSet of Mapping entries
  - except RecoTau/plugins/RecoTauDiscriminantFromDiscriminator.cc
    which is configured in RecoTau/python/RecoTauDiscriminantConfiguration.py
    and I cannot find if it is used anywhere -- probably it is a test
  - other files
       RecoTau/plugins/PATTauDiscriminantCutMultiplexer.cc:    if ( mappingEntry->existsAs<std::string>("cut") ) {}
       RecoTau/plugins/PFTauPrimaryVertexProducer.cc:          if ( pset.existsAs<std::string>("selectionCut") ) {}
       RecoTau/plugins/RecoTauDiscriminantCutMultiplexer.cc:   if ( mappingEntry->existsAs<double>("cut") ) {}
    define a cut with it -- either by some double value, or by a Formula built from a string
    I glanced through configs: they both seem to be used only as double
  - in principle it would be nice to clean up this ambiguity of interface in these MapEntries
    leaving it be for now

And moved verbosity to cfi file of the following plugin:
RecoTauTag/RecoTau/python/PFRecoTauEnergyAlgorithmPlugin_cfi.py
The verbosity was not in cfi because it was set in
RecoTauTag/RecoTau/python/RecoTauCombinatoricProducer_cfi.py
where the plugin is actually used (it is the only place).
It is not Analyzer or Producer or Filter, but a "general" plugin with the parameter set passed to it directly.
Thus reloading the same parameter twice caused a "configuration read error".
Since it is used only in RecoTau Combinatoric production,
maybe it was better to leave verbosity together with the rest parameters -- closer to where they are set.
But the reasoning is to keep the defaults in _cfi file,
to have .cc and _cfi.py files correspond one with the other.
Maybe it's not the best approach, since the plugin is used in such a way.
With this setup you actually _cannot_overwrite_ the default value of verbosity in that PSet() parameter in CombinatoricProducer.
(It is lind of not default but permanent value for the run.)
And a person who runs and debugs the CombinatoricProducer has to change the verbosity directly in this plugin:

    pfTauEnergyAlgorithmPlugin.verbosity = cms.int32(2)

-- before passing it to the PSet().

Other parameters are left where they are set in production:
   RecoTauTag/RecoTau/python/RecoTauCombinatoricProducer_cfi.py
they are
   name = cms.string("tau_en_reconstruction"),
   plugin = cms.string("PFRecoTauEnergyAlgorithmPlugin")
which is, surely, not set in .cc or _cfi.py files.

	modified:   RecoTauTag/RecoTau/plugins/PFRecoTauEnergyAlgorithmPlugin.cc
	modified:   RecoTauTag/RecoTau/plugins/RecoTauDiscriminantCutMultiplexer.cc
	modified:   RecoTauTag/RecoTau/python/PFRecoTauEnergyAlgorithmPlugin_cfi.py
	modified:   RecoTauTag/RecoTau/python/RecoTauCombinatoricProducer_cfi.py