Correction for the default inputFileName parameter (MVA input file)
Is MVA input is not loaded from DB it is taken from the inputFileName.
This variable is passed as cms.FileInPath("<file>"),
which requires the "<file>" to exists at run time before running the plugins.
ifExists() surpasses this feature of EDM plugins system.

In order to keep inputFileName in config defaults and not have ifExists for it,
I added a dummy file in RecoTauTag/RecoTau/test/ directory.

As far as I can see, this parameter is not that important since MVA is always loaded from DB.

	modified:   RecoTauTag/RecoTau/python/PATTauDiscriminantCutMultiplexer_cfi.py
	modified:   RecoTauTag/RecoTau/python/PATTauDiscriminationByMVAIsolationRun2_cff.py
	modified:   RecoTauTag/RecoTau/python/PFRecoTauDiscriminationAgainstMuonMVA_cfi.py
	modified:   RecoTauTag/RecoTau/python/PFRecoTauDiscriminationByMVAIsolation2_cff.py
	modified:   RecoTauTag/RecoTau/python/PFRecoTauDiscriminationByMVAIsolationRun2_cff.py
	modified:   RecoTauTag/RecoTau/python/RecoTauDiscriminantCutMultiplexer_cfi.py
	new file:   RecoTauTag/RecoTau/test/dummyMVAinputFile