Clarified optional param-s in RecoTauTagCleaner
Checked the parameter set validatior to contain all possible parameters
found in the PSets in:
RecoTauTag/RecoTau/python/RecoTauCleanerPlugins.py

Removed the defaults from the validator -- they were generated from one
of PSets in the vector in the cfi file. Also made all not common
parameters to be optional. Changed the naming for clarity and commented
the defined validation description with the default cleaners PSet
vector.

	modified:   RecoTauTag/RecoTau/plugins/RecoTauCleaner.cc